The Blood Pact, A new Slayer Master/Area, Varrock/Draynor progress
- Killing a monster or boss can now progress quest stages (and quests can require killing any arbitrary amount of them)
- Fixed scrolling in the monster and boss collection logs, also added drop quantities to them (already implemented for clue drops)
- Quest Added: The Blood Pact. Xenia, in the Lumbridge Graveyard, saw some Zamorakian cultists dragging a bound woman into the catacombs. Go help her stop them!
- New Area: Lumbridge Catacombs, a low level training area with decent drops.
- New Slayer Master: Jacquelyn, in the Lumbridge Castle Bailey. She is for low level characters, and assigns tasks that are completable in or near Lumbridge (often in the Catacombs).
- New Area: A lot of Varrock, at least some of the key places (plaza for teleports, main passage through the city, sword shop, Aubury's Rune Shop, Horvik's Armor Shop, Lowe's Archery Emporium,
- Bosses can now poison, either with their default attack or only with specials.
- Slayer Equipment Shop implemented, even though basically none of it except the slayer gem and salt is necessary yet.
- Slayer Point Shop implemented. Extra monster assignments all pointless right now, but you can block/extend/prefer/cancel/store tasks. Also there are some items to buy.
- Finally actually fixed the bank / ground item lists being too short or too long randomly
- Shops with a lot of items can now be scrolled up and down
- Added carapace as a tier 4 ranged armor material, dropped by warped cockroaches. Snakeskin moved up to tier 4.5 because it's harder to get.
- Also added ranged shields, made from two bits of material. Mainly useful for use with crossbows, since they're one-handed.
- Slightly changed how item consumption on use works, and hopefully fixed an issue where they were being consumed wrong. New system also allows for items that have charges before running out and optionally destroys them when out of charge (or returns an item, for example an empty vial from a potion)
- New Area: Draynor Village. Includes the market, jail, crossroads, farm, bank, Ned's house
- The Farming Cape is now accessible, purchase it from Martin the Master Gardened in Draynor Market for 99,000 coins if you have 99 Farming.

Locations: 181 (135 + 46)
Items: 1758 (1674 + 84)
Use Recipes: 258 (254 + 4)
Craft Recipes: 393 (378 + 15)
Processing Recipes: 225 (219 + 6)
NPCs: 88 (62 + 26)
Monsters: 61 (36 + 25)
Clue Steps: (10t / 13b / 24e / 8m / 1h)
Quests: 6 (5 + 1)

+++ DataTypes/MaterialDef.cs
@@ -6,15 +6,15 @@
         public int B = 255;
         public int A = 255;
 
-        public int Tier = 1;
+        public double Tier = 1;
         public int Level = 1;
 
         public int CostMultiplier = 1;
         public string Descriptor = "";
 
-        public MaterialDef(string name, Color c, int tier, int lv, int cost, string desc) : this(name, c.R, c.G, c.B, c.A, tier, lv, cost, desc) { }
+        public MaterialDef(string name, Color c, double tier, int lv, int cost, string desc) : this(name, c.R, c.G, c.B, c.A, tier, lv, cost, desc) { }
 
-        public MaterialDef(string name, int r, int g, int b, int a, int tier, int lv, int cost, string desc) {
+        public MaterialDef(string name, int r, int g, int b, int a, double tier, int lv, int cost, string desc) {
             Name = name;
             R = r;
             G = g;
